+++ rmf_fleet_adapter/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 2.1.0 (2022-XX-YY)
 ------------------
+* Add APIs for cancelling and killing tasks from the `RobotUpdateHandle`: [#205](https://github.com/open-rmf/rmf_ros2/pull/205)
 * Add a WaitUntil event and use it for ResponsiveWait: [#199](https://github.com/open-rmf/rmf_ros2/pull/199)
 
 2.0.0 (2022-03-18)